Resolve issue #1 - update Main WIndow.livecode
Resolve issue #1 - The behavior stack for the settings card was not loading when the project was opened.   This was because the StackFiles for the stack 'Main WIndow.livecode' had an incorrect file path to the behaviour script file for the settings card  stack 'Settings Card Behavior.livecodescript'.
To resolve this I deleted the incorrect reference to that stack file and added a reference to that stack file that had the correct file path.