Add a place holder file called .gitkeep to empty directories in the /app directory
Git will not track empty directories.

The Levure /app directory has folders that are part of the app structure that hold specific parts of the app's code in script only stacks.  To make it easier for users to work with the Levure framework and use these stacks without having to create them I have added a place holder file called .gitkeep to each of them to ensure that their enclosing directories will be tracked.

Now when users clone the directory they will get the full application frame work structure